create, update, delete, get by user id and get by address id done

+++ backend/src/main/java/com/example/entity/Address.java
@@ -58,7 +58,7 @@ public class Address {
 		this.post_code = post_code;
 	}
 
-	public long getAddress_id() {
+	public int getAddress_id() {
 		return addressID;
 	}
